# src/server/app.ts
import fastify from 'fastify';
import { z } from 'zod';

import GoogleMapsClient, {
  PlaceAutocompleteMatch,
} from '../clients/GoogleMapsClient';
import { handleServerError } from './errors';

const app = fastify({
  logger: true,
  disableRequestLogging: process.env.NODE_ENV === 'test',
});

const api = {
  googleMaps: new GoogleMapsClient(),
};

interface RestaurantAddress {
  latitude?: number;
  longitude?: number;
  formattedAddress?: string;
}

interface Restaurant {
  id: string;
  name?: string;
  rating?: number;
  location: RestaurantAddress;
}

const searchRestaurantsSchema = z.object({
  query: z.string().min(1),
});

app.get('/places/restaurants', async (request, reply) => {
  const { query } = searchRestaurantsSchema.parse(request.query);

  const fullQuery = `restaurantes em ${query}`;

  const places = await api.googleMaps.searchPlaces(fullQuery, {
    type: 'restaurant',
  });

  const restaurants = places
    .map((place) => {
      const restaurant: Restaurant = {
        id: place.place_id,
        name: place.name,
        rating: place.rating,
        location: {
          latitude: place.geometry?.location.lat,
          longitude: place.geometry?.location.lng,
          formattedAddress: place.formatted_address,
        },
      };

      return restaurant;
    })
    .sort((restaurant, otherRestaurant) => {
      const rating = restaurant.rating ?? 0;
      const otherRating = otherRestaurant.rating ?? 0;
      return otherRating - rating;
    });

  return reply.status(200).send(restaurants);
});

interface PlaceAutocompleteSuggestion {
  text: string;
  formattedText: string;
}

const autocompleteSchema = z.object({
  query: z.string().min(1),
});

/**
 * Esta função formata um texto acrescentando '**' em volta das correspondências.
 *
 * @example
 *  formatAutocompleteText(
 *    'Museu de Arte',
 *    [{ offset: 0, length: 3 }, { offset: 9, length: 1 }]
 *  ); // '**Muse**u de **A**rte'
 */
function formatAutocompleteText(
  text: string,
  matches: PlaceAutocompleteMatch[],
) {
  const textAsArray = text.split('');

  for (const match of matches) {
    const startIndex = match.offset;
    const matchStartCharacter = textAsArray[startIndex];
    textAsArray[match.offset] = `**${matchStartCharacter}`;

    const endIndex = startIndex + match.length - 1;
    const matchEndCharacter = textAsArray[endIndex];
    textAsArray[endIndex] = `${matchEndCharacter}**`;
  }

  return textAsArray.join('');
}

app.get('/places/autocomplete', async (request, reply) => {
  const { query } = autocompleteSchema.parse(request.query);

  const suggestions = await api.googleMaps.autocompletePlacesSearch(query);

  const autocompleteSuggestions = suggestions.map(
    (suggestion): PlaceAutocompleteSuggestion => ({
      text: suggestion.description,
      formattedText: formatAutocompleteText(
        suggestion.description,
        suggestion.matched_substrings,
      ),
    }),
  );

  return reply.status(200).send(autocompleteSuggestions);
});

app.setErrorHandler(handleServerError);

export default app;